Move out from galaxy-util package a file that requires ``galaxy.files``
Also fix commands for `pyupgrade` Makefile target.
