+++ lib/galaxy/files/sources/util.py
